bluesky: Fix "to_json" conversion for DotDict instances.
As I understand it, dict() should work in this case, and this is
likely a bug in the atproto library. The type in question was
app.bsky.feed.threadgate (guessing this has to do with the new
reply-limiting feature, and my atproto module hasn't caught up
yet). Unsure what's really causing this, and I lack the motivation
to find out. The "to_dict" method of DotDict works, so use that.